Added new interface to define troops.
Troops can now be defined via a lvl and type:
- lvl dictate skills (0-4)
- type (inf, missile) dictate attributes (horse, shield, etc.)

This allows to tweak the game balance without the need to change
every troop. By tweaking the definition of a type or skills
of a level, the troops are consistently modified.

For now only the saxon main troops were converted to new interface.